+++ apps/restconf/restconf_stream.h
@@ -42,9 +42,10 @@
  */
 int api_path_is_stream(clixon_handle h);
 int restconf_subscription(clixon_handle h, void *req, char *name, cvec *qvec, int pretty,  restconf_media media_out, int *sp);
-int stream_child_free(clixon_handle h, int pid);
-int stream_child_freeall(clixon_handle h);
-int stream_close(clixon_handle h, void *req);
 int api_stream(clixon_handle h, void *req, cvec *qvec, int timeout, int *finish);
+int stream_sockets_setup(clixon_handle h, void *req, int timeout, int besock, int *finish);
+int stream_close(clixon_handle h, void *req); // only native
+int stream_child_free(clixon_handle h, int pid); // only fcgi
+int stream_child_freeall(clixon_handle h); // only fcgi
 
 #endif /* _RESTCONF_STREAM_H_ */