+++ include/clixon_custom.h
@@ -212,10 +212,6 @@
  */
 #undef USE_SHA256
 
-/*! Restconf native stream support
- */
-#define RESTCONF_NATIVE_STREAM
-
 /*! Temporary comparison of xyanglibs, should be removed asap
  */
 #define YANG_SCHEMA_CMP_KLUDGE